refactor(calendar): extract event aggregation into calendar_service
The GET /calendar/events route body (~213 LOC: enrollment scan, 5 ORM
queries, cv localization, sort + cap) lived in the route function, and
calendar_ical.py cross-imported it from the sibling ROUTE module,
fabricating a throwaway Response() just to satisfy the route signature.

- Add app/services/calendar_service.py with build_calendar_events(db,
  *, user, course_id, limit, display_locale) — pure domain function,
  no Request/Response/Depends; body moved verbatim (same query
  patterns, same fallback chains, same sort/cap semantics).
- calendar.py route becomes a thin wrapper: parse params, set
  Vary: Accept-Language, normalize locale, delegate to the service.
- calendar_ical.py serve_feed calls the service directly; the
  fabricated inner_response hack is deleted.
- test_calendar_ical.py: rename the two monkeypatch targets to the
  new service name (stubs are **kwargs, call site is all-keyword).

Behavior-preserving: response shapes, headers, error envelopes, and
query patterns unchanged. Full suite: 1501 passed before and after.
